# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk sistem pengelolaan museum",
        "title": "API Pengelolaan Gudang Koleksi Museum",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ghaidafasya24",
            "email": "714220031@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "inventorymuseum-de54c3e9b901.herokuapp.com",
    "basePath": "/api",
    "paths": {
        "/kategori": {
            "get": {
                "description": "Mengambil semua data kategori koleksi",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Kategori"
                ],
                "summary": "Get All Kategori",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data kategori museum menggunakan form-data (wajib token)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Kategori"
                ],
                "summary": "Insert Kategori",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama kategori",
                        "name": "nama_kategori",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deskripsi kategori",
                        "name": "deskripsi",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/kategori/{id}": {
            "get": {
                "description": "Mengambil satu data kategori koleksi berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Kategori"
                ],
                "summary": "Get Kategori by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kategori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah data kategori berdasarkan ID. Endpoint ini memerlukan autentikasi JWT Bearer dan menggunakan form-data.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Kategori"
                ],
                "summary": "Update Kategori",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kategori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nama kategori",
                        "name": "nama_kategori",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deskripsi kategori",
                        "name": "deskripsi",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil diperbarui",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data kategori berdasarkan ID (wajib autentikasi JWT Bearer)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Kategori"
                ],
                "summary": "Delete Kategori",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID kategori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kategori berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003cJWT Token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}